+++ app/src/main/java/ru/androidlearning/moviesearch/view/MainActivity.kt
@@ -11,7 +11,7 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.main_activity)
         if (savedInstanceState == null) {
             supportFragmentManager.beginTransaction()
-                    .replace(R.id.container, MainFragment.newInstance())
+                    .replace(R.id.container, MovieDetailFragment.newInstance())
                     .commitNow()
         }
     }
