+++ app/src/main/res/layout/main_fragment.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:theme="@style/textCommonStyle"
-    tools:context=".view.MainFragment">
+    tools:context=".view.MovieDetailFragment">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
@@ -19,7 +19,7 @@
             android:scrollbars="vertical">
 
             <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/moveName"
+                android:id="@+id/movie_name"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="4dp"
@@ -39,10 +39,10 @@
                 android:layout_marginTop="4dp"
                 android:src="@drawable/pulp_fiction"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/moveName" />
+                app:layout_constraintTop_toBottomOf="@+id/movie_name" />
 
             <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/moveGenre"
+                android:id="@+id/movie_genre"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="4dp"
@@ -52,7 +52,7 @@
                 app:layout_constraintTop_toTopOf="@+id/movePoster" />
 
             <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/moveRating"
+                android:id="@+id/movie_rating"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="4dp"
@@ -60,10 +60,10 @@
                 android:layout_marginEnd="4dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@+id/movePoster"
-                app:layout_constraintTop_toBottomOf="@+id/moveGenre" />
+                app:layout_constraintTop_toBottomOf="@+id/movie_genre" />
 
             <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/moveDuration"
+                android:id="@+id/movie_duration"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="4dp"
@@ -71,10 +71,10 @@
                 android:layout_marginEnd="4dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@+id/movePoster"
-                app:layout_constraintTop_toBottomOf="@+id/moveRating" />
+                app:layout_constraintTop_toBottomOf="@+id/movie_rating" />
 
             <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/moveReleaseDate"
+                android:id="@+id/movie_release_date"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="4dp"
@@ -82,10 +82,10 @@
                 android:layout_marginEnd="4dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@+id/movePoster"
-                app:layout_constraintTop_toBottomOf="@+id/moveDuration" />
+                app:layout_constraintTop_toBottomOf="@+id/movie_duration" />
 
             <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/moveDescription"
+                android:id="@+id/movie_description"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="4dp"
