+++ app/src/main/res/values/styles.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
     <style name="textCommonStyle">
         <item name="android:textColor">@color/black</item>
         <item name="android:textSize">16sp</item>
